splitting views and models into several files
