+++ gibbs/models/__init__.py
@@ -1,0 +1,3 @@
+from .reaction import Reaction, CompoundWithCoeff
+from .compound import Compound, CommonName, ValueSource, Specie, SpeciesGroup, \
+                      Reactant, StoredReaction, Enzyme
