+++ views/views.py
@@ -1,55 +1,19 @@
-import base64
 import logging
 import json
 import io
 import os
 from django.http import HttpResponse, HttpResponseBadRequest
-from django.views.decorators.csrf import csrf_exempt
 from django.shortcuts import render
 from django.http import Http404
 from django.template.context_processors import csrf
 from django.apps import apps
-from settings import STATIC_URL
-from gibbs import reaction, conditions, constants, service_config, \
+from gibbs import conditions, constants, service_config, \
                   pathway_result_page
-from gibbs.forms import CompoundForm, EnzymeForm, ReactionForm, \
-                        ReactionGraphForm, SearchForm, SuggestForm, \
+from gibbs.forms import EnzymeForm, SearchForm, SuggestForm, \
                         AnalyzePathwayModelForm, BuildPathwayModelForm
 from pathways import ParsedPathway, PathwayParseError
 import util.django_utils
-
-NO_STRUCTURE_THUMBNAIL = os.path.join([STATIC_URL, 'images',
-                                       'structure_not_available.png'])
-
-
-def MainPage(request):
-    """Renders the landing page."""
-    return render(request, 'main.html', {})
-
-
-def AboutPage(request):
-    """Renders the about page."""
-    return render(request, 'about.html', {})
-
-
-def FAQPage(request):
-    """Renders the FAQ page."""
-    return render(request, 'faq.html', {})
-
-
-def WhatsNewPage(request):
-    """Renders the FAQ page."""
-    return render(request, 'new_in_2_0.html', {})
-
-
-def CitePage(request):
-    """Renders the FAQ page."""
-    return render(request, 'cite.html', {})
-
-
-def ClassicReactions(request):
-    """Renders the classic reactions page."""
-    return render(request, 'classic_reactions.html', {})
+from gibbs.models import Reaction
 
 
 def DownloadPage(request):
@@ -66,61 +30,6 @@ def RefsPage(request):
     return render(request, 'data_refs.html', template_data)
 
 
-def Robots(request):
-    """Renders robots.txt."""
-    return render(request, 'robots.txt', {})
-
-
-@csrf_exempt
-def CompoundImage(request):
-    data = request.GET
-    compound_id = data.get('compoundId', None)
-    if compound_id is None:
-        return HttpResponseBadRequest('No request data.')
-    compounds = apps.get_model('gibbs.Compound').objects.filter(
-        kegg_id=compound_id)
-    if not compounds:
-        return HttpResponseBadRequest('No such compound.')
-    compound = compounds[0]
-    if not compound.thumbnail or compound.thumbnail == 'error':
-        image_data = open(NO_STRUCTURE_THUMBNAIL, 'r').read()
-    else:
-        image_data = base64.decodestring(compound.thumbnail)
-    return HttpResponse(image_data, content_type='image/png')
-
-
-def CompoundPage(request):
-    """Renders a page for a particular compound."""
-    form = CompoundForm(request.GET)
-    if not form.is_valid():
-        logging.error(form.errors)
-        raise Http404
-    aq_params = conditions.AqueousParams.FromForm(form, request.COOKIES)
-    rxn = reaction.Reaction.FromForm(form, aq_params)
-    if len(rxn.reactants) != 1:
-        logging.error('There must be only 1 reactant in a "compound" page')
-        raise Http404
-
-    compound = rxn.reactants[0].compound
-    logging.info('Submit = ' + form.cleaned_submit)
-
-    if form.cleaned_submit == 'Reset':
-        logging.info('resetting conditions')
-        rxn.aq_params = conditions.AqueousParams()
-        rxn.ResetConcentrations()
-
-    compound.StashTransformedSpeciesEnergies(rxn.aq_params)
-
-    query = compound.FirstName()
-    template_data = rxn.GetTemplateData(query)
-    template_data.update({'compound': compound,
-                          'alberty_link': compound.GetNewPriorityLink(99),
-                          'cc_link': compound.GetNewPriorityLink(1)})
-    response = render(request, 'compound_page.html', template_data)
-    rxn.aq_params.SetCookies(response)
-    return response
-
-
 def EnzymePage(request):
     """Renders a page for a particular enzyme."""
     form = EnzymeForm(request.GET)
@@ -132,65 +41,6 @@ def EnzymePage(request):
     template_data = {'is_superuser': util.django_utils.IsSuperUser(request),
                      'enzyme': enz}
     return render(request, 'enzyme_page.html', template_data)
-
-_REACTION_TEMPLATES_BY_SUBMIT = {'': 'reaction_page.html',
-                                 'Update': 'reaction_page.html',
-                                 'Save': 'print_reaction.html',
-                                 'Reverse': 'reaction_page.html',
-                                 'Reset': 'reaction_page.html'}
-
-
-def ReactionPage(request):
-    """Renders a page for a particular reaction."""
-    form = ReactionForm(request.GET)
-    if not form.is_valid():
-        logging.error(form.errors)
-        return HttpResponseBadRequest('Invalid reaction form.')
-    aq_params = conditions.AqueousParams.FromForm(form, request.COOKIES)
-    rxn = reaction.Reaction.FromForm(form, aq_params)
-
-    if form.cleaned_submit == 'Reverse':
-        rxn.SwapSides()
-    elif form.cleaned_submit == 'Reset':
-        rxn.aq_params = conditions.AqueousParams()
-        rxn.ResetConcentrations()
-    query = rxn.GetQueryString()
-    # Render the template.
-    if form.cleaned_submit not in _REACTION_TEMPLATES_BY_SUBMIT:
-        logging.error('Unknown submit term for reaction page: ' +
-                      form.cleaned_submit)
-        raise Http404
-    template_name = _REACTION_TEMPLATES_BY_SUBMIT[form.cleaned_submit]
-    response = render(request, template_name, rxn.GetTemplateData(query))
-    rxn.aq_params.SetCookies(response)
-    return response
-
-
-def ReactionGraph(request):
-    """Renders the graph page."""
-    form = ReactionGraphForm(request.GET)
-    if not form.is_valid():
-        logging.error(form.errors)
-        raise Http404
-
-    mode = 'varyPh'
-    if form.cleaned_vary_is:
-        mode = 'varyIs'
-
-    aq_params = conditions.AqueousParams.FromForm(form, request.COOKIES)
-    rxn = reaction.Reaction.FromForm(form, aq_params)
-
-    compound_list = [c.GetCompoundList() for c in rxn.reactants]
-    coeff_list = [c.coeff for c in rxn.reactants]
-    concentration_list = [c.phase.Value() for c in rxn.reactants]
-    template_data = {'compound_data': json.dumps(compound_list),
-                     'coeff_data': json.dumps(coeff_list),
-                     'concentration_list': json.dumps(concentration_list),
-                     'mode': mode,
-                     'reaction': rxn,
-                     'query': rxn.GetQueryString()}
-    template_data.update(rxn.aq_params.GetTemplateData())
-    return render(request, 'reaction_graph.html', template_data)
 
 
 def ResultsPage(request):
@@ -222,7 +72,7 @@ def ResultsPage(request):
 
         logging.debug('Generating a reaction from the matched KEGG IDs')
         aq_params = conditions.AqueousParams.FromForm(form, request.COOKIES)
-        rxn = reaction.Reaction.FromIds(best_reaction, aq_params)
+        rxn = Reaction.FromIds(best_reaction, aq_params)
         response = render(request, 'reaction_page.html',
                           rxn.GetTemplateData(query))
         return response
